UPDATE a few features
